Renaming FilePreview to AttachmentView
RISDEV-3711
