+++ frontend/src/routes/kitchensink/index/originalFileSidePanel.vue
@@ -1,7 +1,7 @@
 <script lang="ts" setup>
 import { ref } from "vue"
 import { useRoute } from "vue-router"
-import OriginalFileSidePanel from "@/components/FilePreview.vue"
+import OriginalFileSidePanel from "@/components/AttachementView.vue"
 import KitchensinkPage from "@/kitchensink/components/KitchensinkPage.vue"
 import KitchensinkStory from "@/kitchensink/components/KitchensinkStory.vue"
